fix carousel, rename addwidget
